abstract class instead of interface/shared function
