+++ lib/dom/ElementContainer.ts
@@ -1,23 +1,12 @@
 import { StagehandContainer } from "./StagehandContainer";
-import { DomChunk } from "./DomChunk";
-import { collectDomChunksShared } from "@/lib/dom/chunkCollector";
 
-export class ElementContainer implements StagehandContainer {
-  constructor(private el: HTMLElement) {}
+export class ElementContainer extends StagehandContainer {
+  constructor(private el: HTMLElement) {
+    super();
+  }
 
   public getRootElement(): HTMLElement {
     return this.el;
-  }
-
-  public async scrollIntoView(element?: HTMLElement): Promise<void> {
-    if (!element) {
-      // Just scroll ourselves to top
-      this.el.scrollTo({ top: 0, behavior: "smooth" });
-    } else {
-      // If we want to ensure `element` is visible within `this.el`:
-      element.scrollIntoView({ behavior: "smooth", block: "nearest" });
-    }
-    await this.waitForScrollEnd();
   }
 
   public getViewportHeight(): number {
@@ -37,6 +26,15 @@ export class ElementContainer implements StagehandContainer {
     await this.waitForScrollEnd();
   }
 
+  public async scrollIntoView(element?: HTMLElement): Promise<void> {
+    if (!element) {
+      this.el.scrollTo({ top: 0, behavior: "smooth" });
+    } else {
+      element.scrollIntoView({ behavior: "smooth", block: "nearest" });
+    }
+    await this.waitForScrollEnd();
+  }
+
   private async waitForScrollEnd(): Promise<void> {
     return new Promise<void>((resolve) => {
       let scrollEndTimer: number;
@@ -50,25 +48,5 @@ export class ElementContainer implements StagehandContainer {
       this.el.addEventListener("scroll", handleScroll, { passive: true });
       handleScroll();
     });
-  }
-
-  /**
-   * Collect multiple DomChunks from startOffset to endOffset in increments of chunkSize.
-   */
-  public async collectDomChunks(
-    startOffset: number,
-    endOffset: number,
-    chunkSize: number,
-    scrollBackToTop: boolean = true,
-    candidateContainer?: HTMLElement,
-  ): Promise<DomChunk[]> {
-    return collectDomChunksShared(
-      this,
-      startOffset,
-      endOffset,
-      chunkSize,
-      scrollBackToTop,
-      candidateContainer,
-    );
   }
 }
